Eliminate usage of f32::{min, max} (for `no_std`)
I was encountering linking errors on a Cortex-M4 without this change:

    rust-lld: error: undefined symbol: fminf
    >>> referenced by f32.rs:460 (/rustc/c20d7eecbc0928b57da8fe30b2ef8528e2bdd5be/src/libcore/num/f32.rs:460)

    rust-lld: error: undefined symbol: fmaxf
    >>> referenced by f32.rs:444 (/rustc/c20d7eecbc0928b57da8fe30b2ef8528e2bdd5be/src/libcore/num/f32.rs:444)

It seems these are only used one place: computing the RGB color channels,
so I extracted a `color_value` method and implemented the min/max logic
by hand.

With this change it links and runs successfully on Cortex-M4 targets.